+++ changelog.md
@@ -29,6 +29,7 @@ Codion Change Log
 - EntityTableExportPanel now supports cyclical foreign keys.
 - EntityTableExportPanel now automatically expands to show selected attributes when initialized.
 - EntityTableExport now cleans up the file if export is cancelled.
+- EntityTableExportPanel, added save and open configuration.
 
 ## 0.18.51
 ### is.codion.swing.common.ui